Break down "test_distgit" into submodules
`distgit` module has a lot of behavior to be described through tests.
Breaking it down into submodules will reduce considerably the amount of scrolling required to reach a particular test case.
