+++ tests/test_distgit/test_image_distgit.py
@@ -1,157 +1,13 @@
-#!/usr/bin/env
-"""
-Test the management of distgit repositories for RPM and image builds
-
-$ python -m doozerlib.distgit_test
-"""
+import re
+import tempfile
 import unittest
+
 import flexmock
 
-import StringIO
-import logging
-import tempfile
-import shutil
-import re
-from datetime import datetime, timedelta
-
-from dockerfile_parse import DockerfileParser
 import distgit
 from model import Model
 
-
-class MockDistgit(object):
-    def __init__(self):
-        self.branch = None
-
-
-class MockContent(object):
-    def __init__(self):
-        self.branch = None
-
-
-class MockConfig(object):
-
-    def __init__(self):
-        self.distgit = MockDistgit()
-        self.content = Model()
-        self.content.source = Model()
-        self.content.source.specfile = "test-dummy.spec"
-
-
-class SimpleMockLock(object):
-
-    def __enter__(self):
-        pass
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        pass
-
-
-class MockRuntime(object):
-
-    def __init__(self, logger):
-        self.branch = None
-        self.distgits_dir = "distgits_dir"
-        self.logger = logger
-        self.mutex = SimpleMockLock()
-        self.missing_pkgs = set()
-
-    def detect_remote_source_branch(self, _):
-        pass
-
-
-class MockMetadata(object):
-
-    def __init__(self, runtime):
-        self.config = MockConfig()
-        self.runtime = runtime
-        self.logger = runtime.logger
-        self.name = "test"
-        self.namespace = "namespace"
-        self.distgit_key = "distgit_key"
-
-    def fetch_cgit_file(self, file):
-        pass
-
-    def get_component_name(self):
-        pass
-
-
-class MockScanner(object):
-
-    def __init__(self):
-        self.matches = []
-        self.files = []
-
-
-class TestDistgit(unittest.TestCase):
-    """
-    Test the methods and functions used to manage and update distgit repos
-    """
-
-    def setUp(self):
-        """
-        Define and provide mock logging for test/response
-        """
-        self.stream = StringIO.StringIO()
-        logging.basicConfig(level=logging.DEBUG, stream=self.stream)
-        self.logger = logging.getLogger()
-        self.logs_dir = tempfile.mkdtemp()
-        self.md = MockMetadata(MockRuntime(self.logger))
-
-    def tearDown(self):
-        """
-        Reset logging for each test.
-        """
-        logging.shutdown()
-        reload(logging)
-        shutil.rmtree(self.logs_dir)
-
-
-class TestGenericDistGit(TestDistgit):
-    def setUp(self):
-        super(TestGenericDistGit, self).setUp()
-        self.dg = distgit.DistGitRepo(self.md, autoclone=False)
-        self.dg.runtime.group_config = Model()
-
-    def test_init(self):
-        """
-        Ensure that init creates the object expected
-        """
-        self.assertIsInstance(self.dg, distgit.DistGitRepo)
-
-    def test_logging(self):
-        """
-        Ensure that logs work
-        """
-        msg = "Hey there!"
-        self.dg.logger.info(msg)
-
-        actual = self.stream.getvalue()
-
-        self.assertIn(msg, actual)
-
-    @unittest.skip("assertion failing, check if desired behavior changed")
-    def test_add_missing_pkgs_succeed(self):
-        md = MockMetadata(MockRuntime(self.logger))
-        d = distgit.ImageDistGitRepo(md, autoclone=False)
-        d._add_missing_pkgs("haproxy")
-
-        self.assertEqual(1, len(d.runtime.missing_pkgs))
-        self.assertIn("test image is missing package haproxy", d.runtime.missing_pkgs)
-
-    def test_distgit_is_recent(self):
-        scan_freshness = self.dg.runtime.group_config.scan_freshness = Model()
-        self.assertFalse(self.dg.release_is_recent("201901020304"))  # not configured
-
-        scan_freshness.release_regex = r'^(....)(..)(..)(..)'
-        scan_freshness.threshold_hours = 24
-        self.assertFalse(self.dg.release_is_recent("2019"))  # no match by regex
-
-        too_soon = datetime.now() - timedelta(hours=4)
-        self.assertTrue(self.dg.release_is_recent(too_soon.strftime('%Y%m%d%H')))
-        too_stale = datetime.now() - timedelta(hours=25)
-        self.assertFalse(self.dg.release_is_recent(too_stale.strftime('%Y%m%d%H')))
+from .support import MockScanner, TestDistgit
 
 
 class TestImageDistGit(TestDistgit):
@@ -298,53 +154,6 @@ class TestImageDistGit(TestDistgit):
         builds = {"spam-a-lot-container": ("v1", "r1")}
         self.assertTrue(self.img_dg._built_or_recent("v1", "r1", builds))
         self.assertIsNone(self.img_dg._built_or_recent("v2", "r1", builds))
-
-
-class TestRPMDistGit(TestDistgit):
-    def setUp(self):
-        super(TestRPMDistGit, self).setUp()
-        self.rpm_dg = distgit.RPMDistGitRepo(self.md, autoclone=False)
-        self.rpm_dg.runtime.group_config = Model()
-
-    def test_pkg_find_in_spec(self):
-        """ Test RPMDistGitRepo._find_in_spec """
-        self.assertEqual("", self.rpm_dg._find_in_spec("spec", "(?mx) nothing", "thing"))
-        self.assertIn("No thing found", self.stream.getvalue())
-        self.assertEqual("SOMEthing", self.rpm_dg._find_in_spec("no\nSOMEthing", "(?imx) ^ (something)", "thing"))
-
-    def test_pkg_distgit_matches_commit(self):
-        """
-        Check the logic for matching a commit from cgit
-        """
-        flexmock(self.md).should_receive("fetch_cgit_file").once().and_raise(Exception(""))
-        self.assertFalse(self.rpm_dg._matches_commit("anything", {}))
-
-        test_file = u"""
-            Version: 42
-            Release: 201901020304.git.1.12spam7%{?dist}
-            %global commit 5p4mm17y5p4m
-        """
-        flexmock(self.md).should_receive("fetch_cgit_file").and_return(test_file)
-        flexmock(self.rpm_dg).should_receive("_built_or_recent").and_return(None)
-        self.assertIsNone(self.rpm_dg._matches_commit("12spam78", {}))  # matches, falls through
-        self.assertIsNone(self.rpm_dg._matches_commit("5p4mm17y5p4m", {}))  # matches, falls through
-        self.assertFalse(self.rpm_dg._matches_commit("11eggs11", {}))  # doesn't match, returns
-
-        self.assertNotIn("No Release: field found", self.stream.getvalue())
-        flexmock(self.md).should_receive("fetch_cgit_file").once().and_return("nothing")
-        self.assertFalse(self.rpm_dg._matches_commit("nothing", {}))
-        self.assertIn("No Release: field found", self.stream.getvalue())
-
-    def test_pkg_build_or_recent(self):
-        flexmock(self.rpm_dg).should_receive("release_is_recent").and_return(None)
-        self.rpm_dg.name = "mypkg"
-
-        builds = dict(mypkg=("v1", "r1"))
-        self.assertTrue(self.rpm_dg._built_or_recent("v1", "r1", builds))
-
-        builds = dict(mypkg=("v1", "r1.el7"))
-        self.assertTrue(self.rpm_dg._built_or_recent("v1", "r1%{?dist}", builds))
-        self.assertIsNone(self.rpm_dg._built_or_recent("v2", "r1", builds))
 
 
 if __name__ == "__main__":
